requirement.txt
aniso8601==8.0.0
APScheduler==3.6.3
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-APScheduler==1.11.0
Flask-Docs==0.1.4
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.1
pycparser==2.20
pycryptodome==3.9.8
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
